requirements.txt
boto==2.49.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.9.2
dnspython==1.16.0
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pkg-resources==0.0.0
pycparser==2.20
pycryptodome==3.9.7
pymongo==3.10.1
pymongocrypt==1.0.0
requests==2.23.0
six==1.14.0
urllib3==1.25.9
uWSGI==2.0.18
Werkzeug==1.0.1
